SERVER-33605 Mobile SE:Disable capped collections and related testing
(cherry picked from commit e361973f0e994d7c5da603cb6436fd96f7180127)

+++ jstests/noPassthroughWithMongod/find_cmd.js
@@ -21,13 +21,16 @@ assert.eq(0, res.cursor.id);
 assert.eq([], res.cursor.firstBatch);
 
 // Ensure find command keeps cursor open if tailing a capped collection.
-coll.drop();
-assert.commandWorked(coll.getDB().createCollection(collname, {capped: true, size: 2048}));
-assert.writeOK(coll.insert({_id: 1}));
-res = coll.runCommand("find", {tailable: true});
-assert.commandWorked(res);
-assert.neq(0, res.cursor.id);
-assert.eq([{_id: 1}], res.cursor.firstBatch);
+// Skip the test with mobile storage engine as it doesn't support capped collections.
+if (jsTest.options().storageEngine !== "mobile") {
+    coll.drop();
+    assert.commandWorked(coll.getDB().createCollection(collname, {capped: true, size: 2048}));
+    assert.writeOK(coll.insert({_id: 1}));
+    res = coll.runCommand("find", {tailable: true});
+    assert.commandWorked(res);
+    assert.neq(0, res.cursor.id);
+    assert.eq([{_id: 1}], res.cursor.firstBatch);
+}
 
 // Multiple batches.
 coll.drop();
